Create `nw.MenuItem` object and make it an instance of `EventEmitter`
